Put leaf tag code into a separate file.
