+++ HierarchicalTags.py
@@ -1,2 +1,3 @@
 # Keep this file short to make maintenance cleaner
 import hierarchical_tags_addon.hierarchical_tags
+import hierarchical_tags_addon.leaf_tags
